requirements.txt
Django==2.2.7
django-admin-rangefilter==0.5.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.7
django-widget-tweaks==1.4.5
et-xmlfile==1.0.1
jdcal==1.4.1
openpyxl==3.0.0
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
